handle profile and header cleanup
