--- src/app/views/authentication/auth-util-components/ProfileButton.styles.ts
@@ -1,18 +1,0 @@
-import { ITheme } from '@fluentui/react'
-
-export const profileButtonStyles = (theme: ITheme) => {
-  return {
-    actionButtonStyles: {
-      root: {
-        height: 50,
-        width: 50,
-        ':hover': {
-          background: `${theme.palette.neutralLight} !important`
-        }
-      },
-      icon: {
-        flex: 1
-      }
-    }
-  }
-}
