+++ src/app/views/authentication/auth-util-components/ProfileButton.tsx
@@ -1,38 +1,43 @@
-import { ActionButton, getId, getTheme, TooltipHost } from '@fluentui/react';
+import { Button, Persona, PersonaProps, Tooltip } from '@fluentui/react-components';
 import { translateMessage } from '../../../utils/translate-messages';
-import Profile from '../profile/Profile';
-import { profileButtonStyles } from './ProfileButton.styles';
+import { useHeaderStyles } from '../../main-header/utils';
+import { ProfileV9 } from '../profile/Profile';
 
+export const PersonaSignIn = (props: Partial<PersonaProps>) => {
+  return (
+    <Persona
+      presence={{ status: 'offline' }}
+      {...props}
+    />
+  );
+};
+
+const SignInButton = ({signIn}: {signIn: ()=> void})=>{
+  const styles = useHeaderStyles()
+  return (
+    <Tooltip
+      content={translateMessage('sign in')}
+      relationship="description">
+      <Button
+        aria-label={translateMessage('sign in')}
+        role='button'
+        appearance='subtle'
+        className={styles.iconButton}
+        icon={<PersonaSignIn/>}
+        onClick={signIn}
+      />
+    </Tooltip>
+  )
+}
 export function showSignInButtonOrProfile(
   tokenPresent: boolean,
-  signIn: Function,
-  signInWithOther: Function
+  signIn: () => void,
+  signInWithOther: ()=>Promise<void>
 ) {
-
-  const currentTheme = getTheme();
-  const { actionButtonStyles } = profileButtonStyles(currentTheme);
-  const signInButton =
-  <TooltipHost
-    content={
-      <div style={{padding:'3px'}}>
-        {translateMessage('sign in')}
-      </div>}
-    id={getId()}
-    calloutProps={{ gapSpace: 0 }}
-  >
-    <ActionButton
-      ariaLabel={translateMessage('sign in')}
-      role='button'
-      iconProps={{ iconName: 'Contact' }}
-      onClick={() => signIn()}
-      styles={actionButtonStyles}
-    />
-  </TooltipHost>
-
   return (
     <>
-      {!tokenPresent && signInButton}
-      {tokenPresent && <Profile signInWithOther={signInWithOther}/>}
+      {!tokenPresent && <SignInButton signIn={signIn}/>}
+      {tokenPresent && <ProfileV9 signInWithOther={signInWithOther} />}
     </>
   );
 }
