+++ src/app/views/layout/Layout.tsx
@@ -9,7 +9,7 @@ import { translateMessage } from '../../utils/translate-messages';
 import { StatusMessagesV9, TermsOfUseMessageV9 } from '../app-sections';
 import Notification from '../common/banners/Notification';
 import PopupsWrapper from '../common/popups/PopupsWrapper';
-import { MainHeaderV9 } from '../main-header/MainHeaderV9';
+import { MainHeaderV9 } from '../main-header/MainHeader';
 import { QueryResponseV9 } from '../query-response';
 import { QueryRunner } from '../query-runner';
 import Request from '../query-runner/request/RequestV9';
