+++ src/app/views/main-header/FeedbackButton.tsx
@@ -1,52 +1,22 @@
-import { getTheme, IButton, IconButton, IIconProps, TooltipHost } from '@fluentui/react';
-import { useRef, useState, useEffect } from 'react';
-import { translateMessage } from '../../utils/translate-messages';
-import FeedbackForm from '../query-runner/request/feedback/FeedbackForm';
-import { ACCOUNT_TYPE } from '../../services/graph-constants';
-import { componentNames, eventTypes, telemetry } from '../../../telemetry';
-import { useAppSelector } from '../../../store';
+import { Button, MenuTrigger, Tooltip } from '@fluentui/react-components'
+import { PersonFeedback20Regular } from '@fluentui/react-icons'
+import { useState } from 'react'
+import { useAppSelector } from '../../../store'
+import { eventTypes, telemetry } from '../../../telemetry'
+import { FEEDBACK_BUTTON } from '../../../telemetry/component-names'
+import { ACCOUNT_TYPE } from '../../services/graph-constants'
+import { translateMessage } from '../../utils/translate-messages'
+import FeedbackForm from '../query-runner/request/feedback/FeedbackForm'
+import { useHeaderStyles } from './utils'
 
-export const FeedbackButton = () => {
+const trackFeedbackButtonEvent = () => {
+  telemetry.trackEvent(eventTypes.BUTTON_CLICK_EVENT, {
+    ComponentName: FEEDBACK_BUTTON
+  });
+}
+const FeedbackButtonV9 = ()=>{
   const [enableSurvey, setEnableSurvey] = useState(false);
-  const user = useAppSelector((state) => state.profile.user)
-
-  const currentTheme = getTheme();
-  const feedbackIcon : IIconProps = {
-    iconName : 'Feedback'
-  }
-  const feedbackTitle = translateMessage('Feedback');
-  const content = <div style={{padding:'3px'}}>{translateMessage('Feedback')}</div>
-
-  const feedbackButtonRef = useRef<IButton>(null)
-  const isFirstRender = useRef(true);
-  useEffect( () => {
-    if (isFirstRender.current) {
-      isFirstRender.current = false;
-      return;
-    }
-    if(!enableSurvey){
-      feedbackButtonRef.current?.focus();
-    }
-  },[enableSurvey])
-
-  const feedbackIconStyles = {
-    root:{
-      height: '50px',
-      width: '50px',
-      marginTop: '-8px',
-      ':hover': {
-        background: `${currentTheme.palette.neutralLight} !important`
-      }
-    }
-  }
-  const calloutProps = {
-    gapSpace: 0
-  };
-  const hostStyles = { root: {
-    display: 'inline-block'
-  }
-  };
-
+  const user = useAppSelector(state=> state.profile.user)
   const activateSurvey = () => {
     setEnableSurvey(true);
     trackFeedbackButtonEvent();
@@ -56,36 +26,19 @@ export const FeedbackButton = () => {
     setEnableSurvey(false);
   }
 
-  const trackFeedbackButtonEvent = () => {
-    telemetry.trackEvent(eventTypes.BUTTON_CLICK_EVENT, {
-      ComponentName: componentNames.FEEDBACK_BUTTON
-    });
-  }
-
-  return (
-    <div>
-      {user?.profileType !== ACCOUNT_TYPE.AAD &&
-      <div>
-        <TooltipHost
-          content={content}
-          calloutProps={calloutProps}
-          styles={hostStyles}
-        >
-          <IconButton onClick={activateSurvey}
-            iconProps={feedbackIcon}
-            ariaDescription={feedbackTitle}
-            ariaLabel={feedbackTitle}
-            styles={feedbackIconStyles}
-            role={'button'}
-            disabled={enableSurvey}
-            componentRef={feedbackButtonRef}
-          />
-        </TooltipHost>
-
-        <FeedbackForm onDismissSurvey={disableSurvey}
-          activated={enableSurvey} onDisableSurvey={disableSurvey} />
-      </div>
-      }
-    </div>
-  )
+  const styles = useHeaderStyles()
+  return (user?.profileType !== ACCOUNT_TYPE.AAD ? (
+    <>
+      <Tooltip content={translateMessage('Feedback')} relationship="description">
+        <Button
+          onClick={activateSurvey}
+          className={styles.iconButton} appearance="subtle" icon={<PersonFeedback20Regular />} />
+      </Tooltip>
+      <FeedbackForm onDismissSurvey={disableSurvey}
+        activated={enableSurvey} onDisableSurvey={disableSurvey} />
+    </>
+  ) : null)
 }
+
+export { FeedbackButtonV9 }
+
