+++ src/app/views/main-header/settings/ThemeChooser.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { RadioGroup, Radio, DialogActions, Button} from '@fluentui/react-components';
 import { useAppDispatch, useAppSelector } from '../../../../store';
 import { componentNames, eventTypes, telemetry } from '../../../../telemetry';
